Contrib module email_registration added.
This module is added because we are not provided the user names for the
users, so this module will extract the user name from their email.
E.g. If email is abc@xyz.com then the user name will be abc. and if
there are two email addresses like abc@xyz.com and abc@zxy.com then it
will create two user name with name abc and abc_1.

reference : innoraft/kabootr #41